chore(@clayui/tabs): update Tabs component stories
